small frontend stylings and moved cards to display
