+++ frontend/src/components/display/cards/groupCard.js
